Move VS Code language service files into their own folder
