+++ vscode/src/language_service/completion.ts
