+++ vscode/src/language_service/hover.ts
